Refactor code into different line decorator for performance
