+++ src/Extension/Classification/Indent.cs
@@ -86,11 +86,19 @@ namespace IndentRainbow.Extension
                 }
             }
 
-            this.decorator = new LineDecorator(
-                this.drawer, this.colorGetter, this.validator)
+            if(OptionsManager.highlightingMode.Get() == HighlightingMode.Alternating)
             {
-                detectErrors = OptionsManager.detectErrors.Get()
-            };
+                this.decorator = new AlternatingLineDecorator(
+                    this.drawer, this.colorGetter, this.validator)
+                {
+                    detectErrors = OptionsManager.detectErrors.Get()
+                };
+            }
+            if(OptionsManager.highlightingMode.Get() == HighlightingMode.Monocolor)
+            {
+
+            }
+
         }
 
         /// <summary>
