+++ src/Logic/IndentRainbow Logic.csproj
@@ -47,11 +47,12 @@
     <Reference Include="System" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Classification\BaseLineDecorator.cs" />
     <Compile Include="Classification\IIndentValidator.cs" />
     <Compile Include="Classification\ILineDecorator.cs" />
     <Compile Include="Classification\IndentValidator.cs" />
     <Compile Include="Parser\LanguageParser.cs" />
-    <Compile Include="Classification\LineDecorator.cs" />
+    <Compile Include="Classification\AlternatingLineDecorator.cs" />
     <Compile Include="Parser\ColorParser.cs" />
     <Compile Include="Drawing\IBackgroundTextIndexDrawer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
