+++ tests/Logic/IndentRainbow Logic Tests.csproj
@@ -43,7 +43,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="Classification\IndentValidatorTests.cs" />
-    <Compile Include="Classification\LineDecoratorTests.cs" />
+    <Compile Include="Classification\AlternatingLineDecoratorTests.cs" />
     <Compile Include="Parser\ColorParserTests.cs" />
     <Compile Include="Parser\LanguageParserTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
